3/25 골드, 보스 점프 수정. -> 상점 골드로 아이템 사기 버그 고치기
-> 상점 골드 버그 고치기

+++ Assets/1_Script/Enemy/Data/EnemyData.cs
@@ -10,5 +10,5 @@ public class EnemyData : ScriptableObject
     public float atkCooldown;
     public float defense;
     public float rewardExp;
-    public float rewardGold;
+    public int rewardGold;
 }
